IJPS-55 Drop change date time and change by user id fields from sitting record. Update and add tests

+++ src/integrationTest/java/uk/gov/hmcts/reform/jps/repository/StatusHistoryRepositoryTest.java
@@ -6,75 +6,69 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.jps.AbstractTest;
 import uk.gov.hmcts.reform.jps.domain.SittingRecord;
 import uk.gov.hmcts.reform.jps.domain.StatusHistory;
+import uk.gov.hmcts.reform.jps.model.JpsRole;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DataJpaTest
 @ActiveProfiles("itest")
-class StatusHistoryRepositoryTest {
+class StatusHistoryRepositoryTest extends AbstractTest {
 
     @Autowired
     private StatusHistoryRepository historyRepository;
     @Autowired
     private SittingRecordRepository recordRepository;
 
-    private StatusHistory persistedStatusHistory;
+    private SittingRecord persistedSittingRecord;
+    private StatusHistory persistedStatusHistoryCreated;
+    private StatusHistory persistedStatusHistoryAmended;
+    private StatusHistory statusHistoryCreated;
+    private StatusHistory statusHistoryAmended;
 
-    private StatusHistory statusHistory;
 
     @BeforeEach
     public void setUp() {
-        SittingRecord sittingRecord = SittingRecord.builder()
-            .sittingDate(LocalDate.now().minusDays(2))
-            .statusId("recorded")
-            .regionId("1")
-            .epimsId("123")
-            .hmctsServiceId("ssc_id")
-            .personalCode("001")
-            .contractTypeId(2L)
-            .am(true)
-            .judgeRoleTypeId("HighCourt")
-            .createdDateTime(LocalDateTime.now())
-            .createdByUserId("jp-recorder")
-            .build();
-
-        statusHistory = StatusHistory.builder()
-            .statusId("recorded")
-            .changeDateTime(LocalDateTime.now())
-            .changeByUserId("jp-recorder")
-            .changeByName("John Doe")
-            .build();
-
-        sittingRecord.addStatusHistory(statusHistory);
-        SittingRecord persistedSittingRecord = recordRepository.save(sittingRecord);
-        persistedStatusHistory = persistedSittingRecord.getStatusHistories().get(0);
+        SittingRecord sittingRecord = createSittingRecord(LocalDate.now().minusDays(2));
+        statusHistoryCreated = createStatusHistory(sittingRecord.getStatusId(),
+                                                          JpsRole.ROLE_RECORDER.name(),
+                                                          "John Doe",
+                                                            sittingRecord);
+        sittingRecord.addStatusHistory(statusHistoryCreated);
+        persistedSittingRecord = recordRepository.save(sittingRecord);
+        persistedStatusHistoryCreated = persistedSittingRecord.getStatusHistories().get(0);
     }
 
     @Test
     void shouldSaveStatusHistory() {
 
-        assertThat(persistedStatusHistory).isNotNull();
-        assertThat(persistedStatusHistory.getId()).isNotNull();
-        assertThat(persistedStatusHistory).isEqualTo(statusHistory);
+        assertThat(persistedStatusHistoryCreated).isNotNull();
+        assertThat(persistedStatusHistoryCreated.getId()).isNotNull();
+        assertThat(persistedStatusHistoryCreated).isEqualTo(statusHistoryCreated);
     }
 
     @Test
     void shouldUpdateStatusHistoryWhenRecordIsPresent() {
 
         Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository
-            .findById(persistedStatusHistory.getId());
+            .findById(persistedStatusHistoryCreated.getId());
         assertThat(optionalSettingHistoryToUpdate).isPresent();
 
-        StatusHistory settingHistoryToUpdate = optionalSettingHistoryToUpdate.get();
-        settingHistoryToUpdate.setChangeDateTime(LocalDateTime.now());
-        settingHistoryToUpdate.setChangeByUserId("jp-submitter");
+        StatusHistory settingHistoryToUpdate = null;
+        if (optionalSettingHistoryToUpdate.isPresent()) {
+            settingHistoryToUpdate = optionalSettingHistoryToUpdate.get();
+            settingHistoryToUpdate.setChangeDateTime(LocalDateTime.now());
+            settingHistoryToUpdate.setChangeByUserId(JpsRole.ROLE_SUBMITTER.name());
+        }
 
         StatusHistory updatedStatusHistory = historyRepository.save(settingHistoryToUpdate);
         assertThat(updatedStatusHistory).isNotNull();
@@ -82,7 +76,7 @@ class StatusHistoryRepositoryTest {
     }
 
     @Test
-    void shouldReturnEmptyWhenHistorydNotFound() {
+    void shouldReturnEmptyWhenHistoryNotFound() {
         Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository.findById(100L);
         assertThat(optionalSettingHistoryToUpdate).isEmpty();
     }
@@ -90,15 +84,56 @@ class StatusHistoryRepositoryTest {
     @Test
     void shouldDeleteSelectedHistory() {
 
-        Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository
-            .findById(persistedStatusHistory.getId());
+        Optional<StatusHistory> optionalSettingHistoryToUpdate =
+            historyRepository.findById(persistedStatusHistoryCreated.getId());
         assertThat(optionalSettingHistoryToUpdate).isPresent();
 
-        StatusHistory settingHistoryToDelete = optionalSettingHistoryToUpdate.get();
-        historyRepository.deleteById(settingHistoryToDelete.getId());
+        StatusHistory settingHistoryToDelete = null;
+        if (optionalSettingHistoryToUpdate.isPresent()) {
+            settingHistoryToDelete = optionalSettingHistoryToUpdate.get();
+            if (null != settingHistoryToDelete && null != settingHistoryToDelete.getId()) {
+                historyRepository.deleteById(settingHistoryToDelete.getId());
+            }
+        }
 
         optionalSettingHistoryToUpdate = historyRepository.findById(settingHistoryToDelete.getId());
         assertThat(optionalSettingHistoryToUpdate).isEmpty();
+    }
+
+    @Test
+    void shouldFindCreatedStatus() {
+
+        statusHistoryAmended = createStatusHistory("amended",
+                                                   JpsRole.ROLE_RECORDER.name(),
+                                                   "Matthew Doe",
+                                                                 persistedSittingRecord);
+        persistedSittingRecord.addStatusHistory(statusHistoryAmended);
+        persistedStatusHistoryAmended = historyRepository.save(statusHistoryAmended);
+        persistedSittingRecord = recordRepository.save(persistedSittingRecord);
+        persistedStatusHistoryCreated = persistedSittingRecord.getFirstStatusHistory();
+
+        StatusHistory statusHistoryFound = historyRepository
+            .findStatusHistoryAsc(persistedSittingRecord.getId()).get(0);
+        assertNotNull(statusHistoryFound, "Could not find any status history.");
+        assertEquals(statusHistoryFound, statusHistoryCreated, "Not the expected CREATED status history!");
+    }
+
+    @Test
+    void shouldFindLatestStatus() {
+
+        statusHistoryAmended = createStatusHistory("amended",
+                                                                 JpsRole.ROLE_RECORDER.name(),
+                                                                 "Matthew Doe",
+                                                                 persistedSittingRecord);
+        persistedSittingRecord.addStatusHistory(statusHistoryAmended);
+        persistedStatusHistoryAmended = historyRepository.save(statusHistoryAmended);
+        persistedSittingRecord = recordRepository.save(persistedSittingRecord);
+        persistedStatusHistoryCreated = persistedSittingRecord.getFirstStatusHistory();
+
+        StatusHistory statusHistoryFound = historyRepository
+            .findStatusHistoryDesc(persistedSittingRecord.getId()).get(0);
+        assertNotNull(statusHistoryFound, "Could not find any status history.");
+        assertEquals(statusHistoryFound, statusHistoryAmended, "Not the expected AMENDED status history!");
     }
 
 }
